change parseInt to parseFloat in NewKegForm.js

+++ src/components/NewKegForm.js
@@ -17,7 +17,7 @@ function NewKegForm(props) {
     props.onNewKegCreation({
       name: event.target.name.value,
       brand: event.target.brand.value,
-      price: parseInt(event.target.price.value),
+      price: parseFloat(event.target.price.value),
       alcoholContent: event.target.alcoholContent.value,
       pints: 124,
       id: v4() });
